Renaming the path
