+++ sdk/typescript-schema/common/model-path.ts
